requirements.txt
pytest
python-dateutil
python-dotenv
requests
